rename module cupy -> _cupy
